Improve the feedback provided by UpgradeTestHelper.
Previously, UpgradeTestHelper would fail immediately when it encountered
an invalid upgrade step. This meant that a test using the helper would
have to be run multiple times to flush out all the errors.

This change enhances the helper to combine all the errors into a single
failure message where possible:
- All public upgrade steps that should be package-visible will be shown
  as a single failure.
- All errors relating to the sequence, UUID, Jira ID, and description
  will be shown as a single failure.